refactor(S1-WIRE-02): move wire types to shared package and close task
Canonical wire types now live in spacemud-shared/types/wire.ts with a
tsconfig.json for standalone strict validation. The server's wire.ts
becomes a re-export barrel so existing imports are unaffected.

Also corrects the server→client message count (7→8) in DEVELOPMENT_PLAN.md,
strips dead declaration/declarationDir config from the shared tsconfig, and
marks S1-WIRE-02 as complete in SPRINTS.md and the spec.
